# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Onboarding Platform API
  version: 1.0.0
  description: Backend API for AI Onboarding & Training platform
servers:
  - url: /api/v1
    description: V1 base path (documentation only)
...
paths:
...
  /api/v1/admin/users:
    get:
      summary: List all users for admin role management
      description: Returns a page of users in the workspace with their current role, for the Admin role-assignment view.
      tags: [ admin ]
      operationId: listAdminUsers
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - name: search
          in: query
          required: false
          description: Case-insensitive search matched against the user's name and email.
          schema: { type: string, maxLength: 200 }
        - name: role
          in: query
          required: false
          description: Restricts results to users with this role.
          schema: { $ref: '#/components/schemas/UserRoleCodeV1' }
        - name: page
          in: query
          required: false
          description: Zero-based page index.
          schema: { type: integer, format: int32, minimum: 0, default: 0 }
        - name: size
          in: query
          required: false
          description: Maximum number of users returned in one response.
          schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AdminUsersListResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
  /api/v1/admin/users/stats:
    get:
      summary: Get admin user stats
      description: Returns workspace-wide user/admin/team-lead counts for the Admin page stats row, independent of the current search/filter/page.
      tags: [ admin ]
      operationId: getAdminUserStats
      security: [ { bearerAuth: [ ] } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AdminUserStatsResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
  /api/v1/admin/users/{id}/role:
    patch:
      summary: Assign user role
      description: Sets a user's role to Admin, Team Lead, or User. Admin only. Existing group leadership and membership records are left untouched; access checks re-evaluate the user's role on every request.
      tags: [ admin ]
      operationId: assignUserRole
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the assignUserRole.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignUserRoleRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserProfileV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/admin/team-leads:
    get:
      summary: Get team lead admin view
      description: Lists team leads, their teams, and assignable users for admin role management.
      tags: [ admin ]
      operationId: getTeamLeadAdminView
      security: [ { bearerAuth: [ ] } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TeamLeadAdminViewV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
    post:
      summary: Promote team lead
      description: Promotes an existing user to the team lead role by email address.
      tags: [ admin ]
      operationId: promoteTeamLead
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        description: Request payload for the promoteTeamLead.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeamLeadEmailRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserProfileV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
    delete:
      summary: Demote team lead
      description: Demotes a team lead back to member by email address.
      tags: [ admin ]
      operationId: demoteTeamLead
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        description: Request payload for the demoteTeamLead.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeamLeadEmailRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserProfileV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
...
  /api/v1/grades:
    get:
      summary: List grades
      description: Returns active grades in display order. When includeInactive is true, returns every grade including deactivated ones and requires grades.manage.
      tags: [ grades ]
      operationId: listGrades
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - name: includeInactive
          in: query
          required: false
          description: When true, includes deactivated grades. Requires grades.manage.
          schema: { type: boolean, default: false }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GradesListResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
    post:
      summary: Create grade
      description: Creates a new grade dictionary value. Requires grades.manage.
      tags: [ grades ]
      operationId: createGrade
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        description: Request payload for the createGrade.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGradeRequestV1' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/GradeResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
  /api/v1/grades/{id}:
    put:
      summary: Update grade
      description: Renames an existing grade. Requires grades.manage.
      tags: [ grades ]
      operationId: updateGrade
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the updateGrade.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateGradeRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GradeResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
    delete:
      summary: Deactivate grade
      description: Deactivates a grade so it no longer appears in active-grade lists. Existing users keep their current grade reference. Requires grades.manage.
      tags: [ grades ]
      operationId: deactivateGrade
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
  /api/v1/grades/{id}/activate:
    post:
      summary: Activate grade
      description: Reactivates a previously deactivated grade so it appears in active-grade lists again. Requires grades.manage.
      tags: [ grades ]
      operationId: activateGrade
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
...
  /api/v1/files/object:
    get:
      summary: Get file object
      description: Redirects to a presigned URL for downloading a stored object by storage key.
      tags: [ files ]
      operationId: getFileObject
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: storageKey, in: query, required: true, schema: { type: string } }
      responses:
        '307': { description: Redirect to presigned URL }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
  /api/v1/files/preview:
    get:
      summary: Get file preview
      description: Returns preview metadata for a stored object by storage key.
      tags: [ files ]
      operationId: getFilePreview
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: storageKey, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/FilePreviewResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
  /api/v1/files/previews:
    post:
      summary: Get file previews
      description: Returns preview metadata for multiple stored objects in one call, resolving up to 100 storage keys.
      tags: [ files ]
      operationId: getFilePreviews
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        description: Storage keys to resolve into signed preview URLs.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FilePreviewsRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/FilePreviewsResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedV1:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
    ForbiddenV1:
      description: Authenticated but lacking permission
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
    NotFoundV1:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
...
  schemas:
    PageInfoV1:
      description: Pagination metadata attached to a paged list response.
      type: object
      additionalProperties: false
      required: [ page, size, totalElements, totalPages, hasNext, hasPrevious ]
      properties:
        page:
          description: Zero-based index of the page returned in this response.
          type: integer
          format: int32
        size:
          description: Maximum number of items requested per page.
          type: integer
          format: int32
        totalElements:
          description: Total number of items matching the query across all pages.
          type: integer
          format: int64
        totalPages:
          description: Total number of pages available for the query.
          type: integer
          format: int32
        hasNext:
          description: Whether a page after this one exists.
          type: boolean
        hasPrevious:
          description: Whether a page before this one exists.
          type: boolean
...
    UserRoleCodeV1:
      description: User role code.
      type: string
      enum: [ admin, teamlead, member ]
...
    ApiErrorV1:
      description: Standard API error payload.
      type: object
      additionalProperties: false
      required: [ code, message, timestamp ]
      properties:
        code:
          description: Machine-readable code value.
          type: string
        message:
          description: Human-readable message text.
          type: string
        timestamp:
          description: Timestamp for the event or error.
          type: string
          format: date-time
        correlationId:
          description: Correlation identifier for tracing.
          type: string
        parameters:
          description: Parameters for Api Error.
          type: array
          items: { $ref: '#/components/schemas/ValidationParameterV1' }
    ValidationParameterV1:
      description: Single validation parameter attached to an API error.
      type: object
      additionalProperties: false
      required: [ code, value ]
      properties:
        code:
          description: Machine-readable code value.
          type: string
        value:
          description: Parameter value.
          type: string
...
    UserSummaryV1:
      description: Compact user profile shown in team and assignment views.
      type: object
      additionalProperties: false
      required: [ id, email, role ]
      properties:
        id:
          description: Unique identifier.
          type: integer
          format: int64
        name:
          description: Display name.
          type: string
        email:
          description: Email address.
          type: string
          format: email
        role:
          description: Role code for the user or member.
          allOf:
            - $ref: '#/components/schemas/UserRoleCodeV1'
        position:
          description: Job position or title.
          type: string
        avatarStorageKey:
          description: Storage key for the user's avatar image.
          type: string
        avatarColor:
          description: Fallback avatar color when no image is set.
          type: string
        gradeId:
          description: Current grade identifier, when a grade has been assigned.
          type: integer
          format: int64
          nullable: true
        gradeCode:
          description: Current grade's stable code, when a grade has been assigned.
          type: string
          nullable: true
        gradeName:
          description: Current grade's display name, when a grade has been assigned.
          type: string
          nullable: true
    UserProfileV1:
      description: User profile wrapper containing the current user summary.
      type: object
      additionalProperties: false
      required: [ user ]
      properties:
        user:
          description: Nested user payload.
          allOf:
            - $ref: '#/components/schemas/UserSummaryV1'
...
    OkResponseV1:
      description: Ok Response payload.
      type: object
      additionalProperties: false
      required: [ ok ]
      properties:
        ok:
          description: Indicates whether the operation succeeded.
          type: boolean
...
    AdminUsersListResponseV1:
      description: Admin Users List Response payload.
      type: object
      additionalProperties: false
      required: [ users, page ]
      properties:
        users:
          description: One page of users in the workspace with their current role.
          type: array
          items: { $ref: '#/components/schemas/UserSummaryV1' }
        page:
          description: Pagination metadata for the users collection.
          allOf:
            - $ref: '#/components/schemas/PageInfoV1'
    AdminUserStatsV1:
      description: Workspace-wide user/admin/team-lead counts, independent of the current search/filter/page.
      type: object
      additionalProperties: false
      required: [ totalUsers, totalAdmins, totalTeamLeads, totalPermissionsEnabled ]
      properties:
        totalUsers:
          description: Total number of users in the workspace.
          type: integer
          format: int32
        totalAdmins:
          description: Number of users with the admin role.
          type: integer
          format: int32
        totalTeamLeads:
          description: Number of users with the team lead role.
          type: integer
          format: int32
        totalPermissionsEnabled:
          description: Total number of enabled effective permission flags across every team lead.
          type: integer
          format: int32
    AdminUserStatsResponseV1:
      description: Admin User Stats Response payload.
      type: object
      additionalProperties: false
      required: [ stats ]
      properties:
        stats:
          description: The aggregate counts.
          allOf:
            - $ref: '#/components/schemas/AdminUserStatsV1'
    AssignUserRoleRequestV1:
      description: Assign User Role Request payload.
      type: object
      additionalProperties: false
      required: [ roleCode ]
      properties:
        roleCode:
          description: Target role for the user.
          allOf:
            - $ref: '#/components/schemas/UserRoleCodeV1'
    TeamLeadEmailRequestV1:
      description: Team Lead Email Request payload.
      type: object
      additionalProperties: false
      required: [ email ]
      properties:
        email:
          description: Email address.
          type: string
          format: email
    TeamV1:
      description: Team payload.
      type: object
      additionalProperties: false
      required: [ members ]
      properties:
        lead:
          description: Lead for Team.
          allOf:
            - { $ref: '#/components/schemas/UserSummaryV1' }
          nullable: true
        members:
          description: Members for Team.
          type: array
          items: { $ref: '#/components/schemas/UserSummaryV1' }
    TeamLeadAdminViewV1:
      description: Team Lead Admin View payload.
      type: object
      additionalProperties: false
      required: [ teams, users ]
      properties:
        teams:
          description: Teams for Team Lead Admin View.
          type: array
          items: { $ref: '#/components/schemas/TeamV1' }
        users:
          description: Users for Team Lead Admin View.
          type: array
          items: { $ref: '#/components/schemas/UserSummaryV1' }
...
    GradeV1:
      description: A configurable grade dictionary entry.
      type: object
      additionalProperties: false
      required: [ id, code, name, displayOrder, isActive ]
      properties:
        id:
          description: Unique identifier.
          type: integer
          format: int64
        code:
          description: Stable internal code, derived once from the name at creation time.
          type: string
        name:
          description: Display name.
          type: string
        displayOrder:
          description: Sort order for listing.
          type: integer
          format: int32
        isActive:
          description: False when the grade has been deactivated.
          type: boolean
    CreateGradeRequestV1:
      description: Create Grade Request payload.
      type: object
      additionalProperties: false
      required: [ name ]
      properties:
        name:
          description: Display name; the internal code is derived from it. 1-100 characters.
          type: string
          minLength: 1
          maxLength: 100
    UpdateGradeRequestV1:
      description: Update Grade Request payload.
      type: object
      additionalProperties: false
      required: [ name ]
      properties:
        name:
          description: New display name. 1-100 characters.
          type: string
          minLength: 1
          maxLength: 100
    GradeResponseV1:
      description: Grade Response payload.
      type: object
      additionalProperties: false
      required: [ grade ]
      properties:
        grade:
          description: The created or updated grade.
          allOf:
            - $ref: '#/components/schemas/GradeV1'
    GradesListResponseV1:
      description: Grades List Response payload.
      type: object
      additionalProperties: false
      required: [ grades ]
      properties:
        grades:
          description: Grades for Grades List Response.
          type: array
          items: { $ref: '#/components/schemas/GradeV1' }
...
    FilePreviewResponseV1:
      description: File Preview Response payload.
      type: object
      additionalProperties: false
      required: [ previewUrl ]
      properties:
        previewUrl:
          description: URL used to preview the stored object.
          type: string
    FilePreviewsRequestV1:
      description: File Previews Request payload.
      type: object
      additionalProperties: false
      required: [ storageKeys ]
      properties:
        storageKeys:
          description: Storage keys to resolve into signed preview URLs.
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 1024 }
    FilePreviewItemV1:
      description: File Preview Item payload.
      type: object
      additionalProperties: false
      required: [ storageKey, previewUrl ]
      properties:
        storageKey:
          description: Original storage key.
          type: string
        previewUrl:
          description: Signed CloudFront or S3 preview URL.
          type: string
    FilePreviewsResponseV1:
      description: File Previews Response payload.
      type: object
      additionalProperties: false
      required: [ previews ]
      properties:
        previews:
          description: Signed preview URLs keyed by storage key.
          type: array
          items: { $ref: '#/components/schemas/FilePreviewItemV1' }
...